requirements.txt
requests==2.18.4
ldap3==2.4
SQLAlchemy==1.1.14
Flask_SQLAlchemy==2.2
pymssql==2.1.3
Flask==0.12.2
PyJWT==1.5.3
